requirements.txt
gradio>=4.38.1
openai>=1.37.0
python-dotenv>=1.0.1
